+++ lib/_generated/clock/clock_ui_state.freezed.dart
@@ -264,178 +264,3 @@ abstract class _ClockUiState implements ClockUiState {
   _$$ClockUiStateImplCopyWith<_$ClockUiStateImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
-
-/// @nodoc
-mixin _$TimelineState {
-  /// 出発時刻
-  DateTime get departure => throw _privateConstructorUsedError;
-
-  /// format: hh:mm発
-  String get departureTime => throw _privateConstructorUsedError;
-
-  /// format: m分後 || h時間m分後
-  String get remainingTime => throw _privateConstructorUsedError;
-
-  @JsonKey(ignore: true)
-  $TimelineStateCopyWith<TimelineState> get copyWith =>
-      throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class $TimelineStateCopyWith<$Res> {
-  factory $TimelineStateCopyWith(
-          TimelineState value, $Res Function(TimelineState) then) =
-      _$TimelineStateCopyWithImpl<$Res, TimelineState>;
-  @useResult
-  $Res call({DateTime departure, String departureTime, String remainingTime});
-}
-
-/// @nodoc
-class _$TimelineStateCopyWithImpl<$Res, $Val extends TimelineState>
-    implements $TimelineStateCopyWith<$Res> {
-  _$TimelineStateCopyWithImpl(this._value, this._then);
-
-  // ignore: unused_field
-  final $Val _value;
-  // ignore: unused_field
-  final $Res Function($Val) _then;
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? departure = null,
-    Object? departureTime = null,
-    Object? remainingTime = null,
-  }) {
-    return _then(_value.copyWith(
-      departure: null == departure
-          ? _value.departure
-          : departure // ignore: cast_nullable_to_non_nullable
-              as DateTime,
-      departureTime: null == departureTime
-          ? _value.departureTime
-          : departureTime // ignore: cast_nullable_to_non_nullable
-              as String,
-      remainingTime: null == remainingTime
-          ? _value.remainingTime
-          : remainingTime // ignore: cast_nullable_to_non_nullable
-              as String,
-    ) as $Val);
-  }
-}
-
-/// @nodoc
-abstract class _$$TimelineStateImplCopyWith<$Res>
-    implements $TimelineStateCopyWith<$Res> {
-  factory _$$TimelineStateImplCopyWith(
-          _$TimelineStateImpl value, $Res Function(_$TimelineStateImpl) then) =
-      __$$TimelineStateImplCopyWithImpl<$Res>;
-  @override
-  @useResult
-  $Res call({DateTime departure, String departureTime, String remainingTime});
-}
-
-/// @nodoc
-class __$$TimelineStateImplCopyWithImpl<$Res>
-    extends _$TimelineStateCopyWithImpl<$Res, _$TimelineStateImpl>
-    implements _$$TimelineStateImplCopyWith<$Res> {
-  __$$TimelineStateImplCopyWithImpl(
-      _$TimelineStateImpl _value, $Res Function(_$TimelineStateImpl) _then)
-      : super(_value, _then);
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? departure = null,
-    Object? departureTime = null,
-    Object? remainingTime = null,
-  }) {
-    return _then(_$TimelineStateImpl(
-      departure: null == departure
-          ? _value.departure
-          : departure // ignore: cast_nullable_to_non_nullable
-              as DateTime,
-      departureTime: null == departureTime
-          ? _value.departureTime
-          : departureTime // ignore: cast_nullable_to_non_nullable
-              as String,
-      remainingTime: null == remainingTime
-          ? _value.remainingTime
-          : remainingTime // ignore: cast_nullable_to_non_nullable
-              as String,
-    ));
-  }
-}
-
-/// @nodoc
-
-class _$TimelineStateImpl implements _TimelineState {
-  _$TimelineStateImpl(
-      {required this.departure,
-      required this.departureTime,
-      required this.remainingTime});
-
-  /// 出発時刻
-  @override
-  final DateTime departure;
-
-  /// format: hh:mm発
-  @override
-  final String departureTime;
-
-  /// format: m分後 || h時間m分後
-  @override
-  final String remainingTime;
-
-  @override
-  String toString() {
-    return 'TimelineState(departure: $departure, departureTime: $departureTime, remainingTime: $remainingTime)';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$TimelineStateImpl &&
-            (identical(other.departure, departure) ||
-                other.departure == departure) &&
-            (identical(other.departureTime, departureTime) ||
-                other.departureTime == departureTime) &&
-            (identical(other.remainingTime, remainingTime) ||
-                other.remainingTime == remainingTime));
-  }
-
-  @override
-  int get hashCode =>
-      Object.hash(runtimeType, departure, departureTime, remainingTime);
-
-  @JsonKey(ignore: true)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$TimelineStateImplCopyWith<_$TimelineStateImpl> get copyWith =>
-      __$$TimelineStateImplCopyWithImpl<_$TimelineStateImpl>(this, _$identity);
-}
-
-abstract class _TimelineState implements TimelineState {
-  factory _TimelineState(
-      {required final DateTime departure,
-      required final String departureTime,
-      required final String remainingTime}) = _$TimelineStateImpl;
-
-  @override
-
-  /// 出発時刻
-  DateTime get departure;
-  @override
-
-  /// format: hh:mm発
-  String get departureTime;
-  @override
-
-  /// format: m分後 || h時間m分後
-  String get remainingTime;
-  @override
-  @JsonKey(ignore: true)
-  _$$TimelineStateImplCopyWith<_$TimelineStateImpl> get copyWith =>
-      throw _privateConstructorUsedError;
-}
